add repl
