+++ lisp/lisp-rs/src/repl.rs
@@ -1,0 +1,34 @@
+use anyhow::Result;
+use rustyline::{Editor, history::FileHistory, error::ReadlineError};
+
+pub fn run() -> Result<()> {
+    let mut rl = Editor::<(), FileHistory>::new()?;
+    let _ = rl.load_history(".mal-history");
+
+    loop {
+        let readline = rl.readline("> ");
+        match readline {
+            Ok(input) => {
+                if !input.trim().is_empty() {
+                    println!("{}", input);
+                    let _ = rl.add_history_entry(&input)?;
+                    let _ = rl.save_history(".lisp-rs-history")?;
+                }
+            },
+            Err(ReadlineError::Interrupted) => {
+                println!("Interrupted.");
+                break;
+            },
+            Err(ReadlineError::Eof) => {
+                println!("Bye!");
+                break;
+            },
+            Err(err) => {
+                eprintln!("Error: {:?}", err);
+                break;
+            },
+        }
+    }
+
+    Ok(())
+}
